backend/requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pypdf
python-dotenv
fastapi
uvicorn
python-multipart
tiktoken
